requirements.txt
streamlit>=1.28.0
ollama>=0.1.0
python-dotenv>=1.0.0
anthropic>=0.7.0
openai>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
